polaris_poc/requirements.txt
nats-py
asyncio-nats-client
pydantic
# langchain
# langchain-google-genai
python-dotenv
asyncio
